Moved GetDevices to devices.go
